backend/src/constants.py
import os

from dotenv import load_dotenv

load_dotenv()


POSTGRES_USER = os.getenv('POSTGRES_USER')
POSTGRES_PASSWORD = os.getenv('POSTGRES_PASSWORD')
POSTGRES_DB = os.getenv('POSTGRES_DB')
DB_HOST = os.getenv('DB_HOST', 'localhost')
DB_PORT = os.getenv('DB_PORT', '5432')
UPDATE_PERIOD = os.getenv('UPDATE_PERIOD')
HTTP_PREFIX = os.getenv('HTTP_PREFIX')

TIMEOUT = 2
