backend/requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioresponses==0.7.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
exceptiongroup==1.2.2
executor==23.2
fastapi==0.112.2
fasteners==0.19
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
humanfriendly==10.0
idna==3.8
iniconfig==2.0.0
magic-filter==1.0.12
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
property-manager==3.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pyreadline3==3.4.1
pytest==8.3.2
pytest-asyncio==0.24.0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uvicorn==0.30.6
verboselogs==1.7
yarl==1.9.4
